requirements.txt
fastapi[standard]==0.115.5
aiohttp[speedups]==3.11.2
PyYAML==6.0.2
openpyxl==3.1.5
aiofiles==24.1.0
aiosqlite==0.20.0
